+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * [FEATURE] Add uname collector for FreeBSD #1239
 * [FEATURE] Add diskstats collector for OpenBSD #1250
 * [CHANGE] Bonding state uses mii_status #1124
+* [FEATURE] Add a flag to enable process metrics for systemd services #1289
 
 ## 0.17.0 / 2018-11-30